+++ Package.swift
@@ -14,6 +14,7 @@ let package = Package(
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
+        .package(url: "https://github.com/Quick/Nimble", .branch("master"))
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -23,6 +24,6 @@ let package = Package(
             dependencies: []),
         .testTarget(
             name: "DataKitTests",
-            dependencies: ["DataKit"]),
+            dependencies: ["DataKit", "Nimble"])
     ]
 )
