+++ Tests/DataKitTests/DataExtASCIIStringTests.swift
@@ -1,22 +1,22 @@
 //
-// swiftlint:disable:previous file_name
 // Created by Arjan Duijzer on 24/06/2018.
 //
 
 @testable import DataKit
+import Nimble
 import XCTest
 
-final class DataASCIIStringTests: XCTestCase {
+final class DataExtASCIIStringTests: XCTestCase {
 
     func testIncompatibleData() {
         let testString = "Héllø, this îs ån ÜTF8 ßtrìnƒ"
         let data = testString.data(using: .utf8)
-        XCTAssertNil(data?.asciiString)
+        expect(data?.asciiString).to(beNil())
     }
 
     func testASCIIData() {
         let data = Data(bytes: [97, 98, 99, 100])
-        XCTAssertEqual("abcd", data.asciiString)
+        expect(data.asciiString).to(equal("abcd"))
     }
 
     static var allTests = [
